New Order Form

+++ WarehouseManagement/BusinesLogic/Order.cs
@@ -8,13 +8,13 @@ namespace BusinesLogic
 {
     public class Order
     {
-        public int Id { get; private set; }
+        public Int64 Id { get; private set; }
         public DateTime Date { get; private set; }
         public List<Part> Parts { get; private set; }
         public string Customer { get; private set; }
         public string Status { get; set; }
 
-        public Order(int id, DateTime date, List<Part> parts, string customer, string status)
+        public Order(Int64 id, DateTime date, List<Part> parts, string customer, string status)
         {
             Id = id;
             Date = date;
